requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2021.5
psycopg2-binary==2.9.4
